+++ src/textTools.js
@@ -118,7 +118,11 @@ TextTools.prototype.widthOfString = function (text, font, fontSize, characterSpa
 	return widthOfString(text, font, fontSize, characterSpacing, fontFeatures);
 };
 
-function splitWords(text, noWrap) {
+function splitWords(text, noWrap, breakAll) {
+	if (breakAll === undefined) {
+		breakAll = false;
+	}
+
 	var results = [];
 
 	if (text === undefined || text === null) {
@@ -137,6 +141,14 @@ function splitWords(text, noWrap) {
 	if (text.length === 0) {
 	 	results.push({ text: '' });
 	 	return results;
+	}
+	if (breakAll) {
+		return text.split('').map(c => {
+			if(c.match(/^\n$|^\r$/)) { // new line
+				return { text: '', lineEnd: true };
+			}
+			return { text: c };
+		});
 	}
 
 	var breaker = new LineBreaker(text);
@@ -217,15 +229,16 @@ function normalizeTextArray(array, styleContextStack) {
 		var style = null;
 		var words;
 
+		var breakAll = getStyleProperty(item || {}, styleContextStack, 'wordBreak', 'normal') === 'break-all';
 		var noWrap = getStyleProperty(item || {}, styleContextStack, 'noWrap', false);
 		if (isObject(item)) {
 			if (item._textRef && item._textRef._textNodeRef.text) {
 				item.text = item._textRef._textNodeRef.text;
 			}
-			words = splitWords(normalizeString(item.text), noWrap);
+			words = splitWords(normalizeString(item.text), noWrap, breakAll);
 			style = copyStyle(item);
 		} else {
-			words = splitWords(normalizeString(item), noWrap);
+			words = splitWords(normalizeString(item), noWrap, breakAll);
 		}
 
 		if (lastWord && words.length) {